Nick: refactor
