+++ apps/api/src/scraper/WebScraper/global.ts
@@ -1,0 +1,1 @@
+export const universalTimeout = 15000;
